+++ tests/conftest.py
@@ -1,25 +1,15 @@
-import asyncio
 import functools
 import logging
 import re
-from asyncio import Event
-from asyncio.base_events import Server
-from contextlib import asynccontextmanager
-import socket
-from typing import Optional, Tuple
 
 import pytest
 
 from rsocket.frame_parser import FrameParser
-from rsocket.helpers import single_transport_provider
-from rsocket.rsocket_base import RSocketBase
-from rsocket.rsocket_client import RSocketClient
-from rsocket.rsocket_server import RSocketServer
-from rsocket.transports.aiohttp_websocket import websocket_client, websocket_handler_factory
-from rsocket.transports.aioquic_transport import rsocket_connect, rsocket_serve
-
-from rsocket.transports.tcp import TransportTCP
-from tests.rsocket.helpers import assert_no_open_streams
+# noinspection PyUnresolvedReferences
+from tests.tools.fixtures_aiohttp import pipe_factory_aiohttp_websocket, aiohttp_raw_server
+from tests.tools.fixtures_aioquic import pipe_factory_quic
+from tests.tools.fixtures_quart import pipe_factory_quart_websocket
+from tests.tools.fixtures_tcp import pipe_factory_tcp
 
 logging.basicConfig(level=logging.DEBUG)
 
@@ -98,210 +88,6 @@ async def pipe_tcp_without_auto_connect(unused_tcp_port):
         yield components
 
 
-@asynccontextmanager
-async def pipe_factory_tcp(unused_tcp_port, client_arguments=None, server_arguments=None, auto_connect_client=True):
-    wait_for_server = Event()
-
-    def session(*connection):
-        nonlocal server
-        server = RSocketServer(TransportTCP(*connection), **(server_arguments or {}))
-        wait_for_server.set()
-
-    async def start():
-        nonlocal service, client
-        service = await asyncio.start_server(session, host, port)
-        connection = await asyncio.open_connection(host, port)
-        nonlocal client_arguments
-        # test_overrides = {'keep_alive_period': timedelta(minutes=20)}
-        client_arguments = client_arguments or {}
-
-        # client_arguments.update(test_overrides)
-
-        client = RSocketClient(single_transport_provider(TransportTCP(*connection)), **(client_arguments or {}))
-
-        if auto_connect_client:
-            await client.connect()
-
-    async def finish():
-        if auto_connect_client:
-            await client.close()
-
-        await server.close()
-
-        service.close()
-
-    service: Optional[Server] = None
-    server: Optional[RSocketServer] = None
-    client: Optional[RSocketClient] = None
-    port = unused_tcp_port
-    host = 'localhost'
-
-    await start()
-
-    async def server_provider():
-        await wait_for_server.wait()
-        return server
-
-    try:
-        if auto_connect_client:
-            await wait_for_server.wait()
-            yield server, client
-        else:
-            yield server_provider, client
-
-        assert_no_open_streams(client, server)
-    finally:
-        await finish()
-
-
 @pytest.fixture
 def frame_parser():
     return FrameParser()
-
-
-@pytest.fixture
-def aiohttp_raw_server(event_loop: asyncio.BaseEventLoop, unused_tcp_port):
-    from aiohttp.test_utils import RawTestServer
-
-    servers = []
-
-    async def go(handler, *args, **kwargs):  # type: ignore[no-untyped-def]
-        server = RawTestServer(handler, port=unused_tcp_port)
-        await server.start_server(**kwargs)
-        servers.append(server)
-        return server
-
-    yield go
-
-    async def finalize() -> None:
-        while servers:
-            await servers.pop().close()
-
-    event_loop.run_until_complete(finalize())
-
-
-@asynccontextmanager
-async def pipe_factory_quic(unused_tcp_port,
-                            client_arguments=None,
-                            server_arguments=None):
-    server: Optional[RSocketBase] = None
-    wait_for_server = Event()
-
-    def store_server(new_server):
-        nonlocal server
-        server = new_server
-        wait_for_server.set()
-
-    quic_server = await rsocket_serve(host='localhost',
-                                      port=unused_tcp_port,
-                                      on_server_create=store_server,
-                                      **(server_arguments or {}))
-
-    # test_overrides = {'keep_alive_period': timedelta(minutes=20)}
-    client_arguments = client_arguments or {}
-    # client_arguments.update(test_overrides)
-    transport = await rsocket_connect('localhost', unused_tcp_port)
-
-    async with RSocketClient(single_transport_provider(transport),
-                             **client_arguments) as client:
-        await wait_for_server.wait()
-        yield server, client
-        await server.close()
-        assert_no_open_streams(client, server)
-
-    quic_server.cancel()
-    await quic_server
-
-
-@asynccontextmanager
-async def pipe_factory_aiohttp_websocket(aiohttp_raw_server, unused_tcp_port, client_arguments=None,
-                                         server_arguments=None):
-    server: Optional[RSocketBase] = None
-    wait_for_server = Event()
-
-    def store_server(new_server):
-        nonlocal server
-        server = new_server
-        wait_for_server.set()
-
-    await aiohttp_raw_server(websocket_handler_factory(on_server_create=store_server, **(server_arguments or {})))
-
-    # test_overrides = {'keep_alive_period': timedelta(minutes=20)}
-    client_arguments = client_arguments or {}
-    # client_arguments.update(test_overrides)
-
-    async with websocket_client('http://localhost:{}'.format(unused_tcp_port),
-                                **client_arguments) as client:
-        await wait_for_server.wait()
-        yield server, client
-        await server.close()
-        assert_no_open_streams(client, server)
-
-
-@asynccontextmanager
-async def pipe_factory_quart_websocket(unused_tcp_port, client_arguments=None, server_arguments=None):
-    from quart import Quart
-    from rsocket.transports.quart_websocket import websocket_handler
-
-    app = Quart(__name__)
-    server: Optional[RSocketBase] = None
-    wait_for_server = Event()
-
-    def store_server(new_server):
-        nonlocal server
-        server = new_server
-        wait_for_server.set()
-
-    @app.websocket("/")
-    async def ws():
-        await websocket_handler(on_server_create=store_server, **(server_arguments or {}))
-        # test_overrides = {'keep_alive_period': timedelta(minutes=20)}
-
-    client_arguments = client_arguments or {}
-    # client_arguments.update(test_overrides)
-    server_task = asyncio.create_task(app.run_task(port=unused_tcp_port))
-    await asyncio.sleep(0.1)
-
-    async with websocket_client('http://localhost:{}'.format(unused_tcp_port),
-                                **client_arguments) as client:
-        await wait_for_server.wait()
-        yield server, client
-        await server.close()
-        assert_no_open_streams(client, server)
-
-    try:
-        server_task.cancel()
-        await server_task
-    except asyncio.CancelledError:
-        pass
-
-
-def generate_openssl_certificate_and_key() -> Tuple[str, bytes]:
-    import random
-    from OpenSSL import crypto
-
-    pkey = crypto.PKey()
-    pkey.generate_key(crypto.TYPE_RSA, 2048)
-
-    x509 = crypto.X509()
-    subject = x509.get_subject()
-    subject.commonName = socket.gethostname()
-    x509.set_issuer(subject)
-    x509.gmtime_adj_notBefore(0)
-    x509.gmtime_adj_notAfter(5 * 365 * 24 * 60 * 60)
-    x509.set_pubkey(pkey)
-    x509.set_serial_number(random.randrange(100000))
-    x509.set_version(2)
-    x509.add_extensions([
-        crypto.X509Extension(b'subjectAltName', False,
-                             ','.join([
-                                 'DNS:%s' % socket.gethostname(),
-                                 'DNS:*.%s' % socket.gethostname(),
-                                 'DNS:localhost',
-                                 'DNS:*.localhost']).encode()),
-        crypto.X509Extension(b"basicConstraints", True, b"CA:false")])
-
-    x509.sign(pkey, 'SHA256')
-
-    return (crypto.dump_certificate(crypto.FILETYPE_PEM, x509),
-            crypto.dump_privatekey(crypto.FILETYPE_PEM, pkey))
